Messages now converted to uppercase
Messages are converted to uppercase and the bot now send the user a private message instead of sending in the group. Cleanup has started and needs to continue.
